modified exception handling

+++ src/main/java/com/chatmate/social/entity/Post.java
@@ -4,13 +4,13 @@ import javax.validation.constraints.NotBlank;
 
 public class Post {
 
-    private Integer postId;
+    private Long postId;
 
     @NotBlank
     private String body;
 
     private long timestamp;
 
-    private Integer userId;
+    private Long userId;
 
 }
